+++ lib/cubix.dart
@@ -5,6 +5,44 @@ import 'dart:async';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+Transformer debounce(Duration duration) {
+  return (context, next) {
+    context.previous?.cancelToken.cancel();
+    context.cancelToken.onCancel(Timer(duration, next).cancel);
+  };
+}
+
+Transformer droppable() {
+  return (context, next) {
+    if (context.count >= 2) {
+      return context.cancelToken.cancel();
+    }
+    next();
+  };
+}
+
+Transformer sequential() {
+  return (context, next) {
+    if (context.previous != null) {
+      context.previous!.onDone(next);
+    } else {
+      next();
+    }
+  };
+}
+
+Transformer throttle(Duration duration) {
+  return (context, next) {
+    final lastTime = context.get<DateTime?>(_Props.throttleLastExecutionTime);
+    final now = DateTime.now();
+    final nextTime = lastTime?.add(duration);
+    if (nextTime == null || nextTime.compareTo(now) <= 0) {
+      context.set(_Props.throttleLastExecutionTime, now);
+      next();
+    }
+  };
+}
+
 Future<T> _neverDone<T>() => Completer<T>().future;
 
 void _noop() {}
@@ -14,6 +52,9 @@ typedef CreateCubix<T extends Cubix> = T Function();
 typedef OnCancel = VoidCallback Function(VoidCallback handler);
 
 typedef ResolveCubix<T extends Cubix> = T Function(BuildContext context);
+
+typedef Transformer = void Function(
+    TransformContext context, VoidCallback next);
 
 class AsyncContext {
   final bool Function() _isCancelled;
@@ -50,6 +91,8 @@ abstract class AsyncCubix<T> extends Cubix<AsyncState<T>> {
   final _onCancelHandlers = <VoidCallback>{};
 
   AsyncCubix(T initialState) : super(AsyncState(initialState));
+
+  T get data => state.data;
 
   Object? get error => state.error;
 
@@ -123,23 +166,35 @@ abstract class AsyncCubix<T> extends Cubix<AsyncState<T>> {
   /// dispatch async action and handle loading state and error state if any
   @protected
   Future<TResult> dispatchAsync<TResult>(
-    Future<TResult> Function(DispatchAsyncContext context) action, {
+    Future<TResult> Function(DispatchAsyncContext<T> context) action, {
+    Function? key,
+    List<Transformer>? transform,
     CancelToken? cancelToken,
   }) {
     if (cancelToken?.cancelled == true) return _neverDone();
+    if (transform != null || key != null) {
+      if (transform == null || key == null) {
+        throw Exception('transform and key cannot be null');
+      }
+      final originAction = action;
+      action = (context) =>
+          this.transform(key, transform, () => originAction(context));
+    }
+
+    _emit(AsyncState(state.data, loading: true));
 
     var token = _token;
-    final context = DispatchAsyncContext(
+    final context = DispatchAsyncContext<T>(
       (T value) {
         if (cancelToken?.cancelled == true) return;
-        _emit(AsyncState(value), token);
+        if (_emit(AsyncState(value), token)) {
+          token = _token;
+        }
       },
       () => _token != token || cancelToken?.cancelled == true,
       _onCancel,
     );
     final completer = Completer<TResult>();
-
-    _emit(AsyncState(state.data, loading: true));
 
     action(context).then(
       (value) {
@@ -268,6 +323,8 @@ class CreateContext {
 abstract class Cubix<T> extends Cubit<T> {
   var _disposed = false;
   final _onDispose = <VoidCallback>{};
+  final _transformData = <Object?, Map<Object?, Object?>>{};
+  final _executing = <Function>[];
   late final DependencyResolver _resolver;
   late final Type _resolvedType;
   Object? _key;
@@ -298,6 +355,74 @@ abstract class Cubix<T> extends Cubit<T> {
   /// this method will run once if sync configs is not enabled
   /// if sync configs is enabled, it will run whenever dependency cubixes are updated
   void onInit() {}
+
+  Future<TResult> transform<TResult>(
+    Function key,
+    List<Transformer> trasnformers,
+    Future<TResult> Function() execute, {
+    CancelToken? cancelToken,
+  }) {
+    cancelToken ??= CancelToken();
+    var done = false;
+    final completer = Completer<TResult>();
+
+    if (trasnformers.isEmpty) {
+      throw Exception('transformers cannot be empty');
+    }
+
+    TransformContext? current;
+    var data = _transformData[key];
+    if (data == null) {
+      data = {};
+      _transformData[key] = data;
+    }
+
+    void onDone() {
+      if (done) return;
+      done = true;
+      if (data![_Props.previousContext] == current) {
+        data.remove(_Props.previousContext);
+      }
+      _executing.remove(key);
+    }
+
+    void next() {
+      execute().then(
+        (result) {
+          if (cancelToken?.cancelled == true) return;
+          completer.complete(result);
+        },
+        onError: (e) {
+          if (cancelToken?.cancelled == true) return;
+          completer.completeError(e);
+        },
+      ).whenComplete(onDone);
+    }
+
+    cancelToken.onCancel(onDone);
+
+    _executing.add(key);
+
+    final context = current = TransformContext(
+      data[_Props.previousContext] as TransformContext?,
+      key,
+      data,
+      cancelToken,
+      completer.future,
+      _executing,
+    );
+    data[_Props.previousContext] = current;
+    final invoker = trasnformers.reversed.fold<VoidCallback>(
+      next,
+      (next, m) => () {
+        if (cancelToken?.cancelled == true) return;
+        m(context, next);
+      },
+    );
+    invoker();
+
+    return completer.future;
+  }
 
   /// remove cubix from provider
   void _remove() {
@@ -533,6 +658,47 @@ class DispatchAsyncContext<T> extends AsyncContext {
     _emit(value);
   }
 }
+
+class TransformContext {
+  final Function key;
+  final Map<Object?, Object?> _data;
+  final CancelToken cancelToken;
+  final Future _future;
+  final List<Function> executing;
+  final TransformContext? previous;
+
+  TransformContext(
+    this.previous,
+    this.key,
+    this._data,
+    this.cancelToken,
+    this._future,
+    this.executing,
+  );
+
+  int get count => executing.where((element) => element == key).length;
+
+  T? get<T extends Object?>(Object? name) {
+    return _data[name] as T?;
+  }
+
+  void onDone(VoidCallback callback) {
+    _future.whenComplete(callback);
+  }
+
+  set(Object? name, Object? value) {
+    _data[name] = value;
+  }
+
+  T tryGet<T extends Object?>(Object? name, T Function() create) {
+    if (_data.containsKey(name)) {
+      return _data[name] as T;
+    }
+    return _data[name] = create();
+  }
+}
+
+enum _Props { previousContext, throttleLastExecutionTime }
 
 class _SyncConfigs {
   final Duration? debounce;